+++ Tests/NSMutableIndexSetBlocksKitTest.m
@@ -18,8 +18,8 @@
 
 - (void)testSelect {
 	NSMutableString *order = [NSMutableString string];
-	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld",index];
+	BKIndexValidationBlock indexValidationBlock = ^(unsigned int index) {
+		[order appendFormat:@"%d",index];
 		BOOL match = index < 3 ? YES : NO; //1,2
 		return match;
 	};
@@ -31,8 +31,8 @@
 
 - (void)testSelectedNone {
 	NSMutableString *order = [NSMutableString string];
-	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld",index];
+	BKIndexValidationBlock indexValidationBlock = ^(unsigned int index) {
+		[order appendFormat:@"%d",index];
 		BOOL match = index == 0 ? YES : NO;
 		return match;
 	};
@@ -43,8 +43,8 @@
 
 - (void)testReject {
 	NSMutableString *order = [NSMutableString string];
-	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld",index];
+	BKIndexValidationBlock indexValidationBlock = ^(unsigned int index) {
+		[order appendFormat:@"%d",index];
 		BOOL match = [[_target objectAtIndex:index] isEqual: @"0"] ? YES : NO;
 		return match;
 	};
@@ -55,8 +55,8 @@
 
 - (void)testRejectedNone {
 	NSMutableString *order = [NSMutableString string];
-	BKIndexValidationBlock indexValidationBlock = ^(NSUInteger index) {
-		[order appendFormat:@"%ld",index];
+	BKIndexValidationBlock indexValidationBlock = ^(unsigned int index) {
+		[order appendFormat:@"%d",index];
 		BOOL match = [[_target objectAtIndex:index] isEqual: @"0"] ? NO : YES;
 		return match;
 	};
